BE - refactor game controller
